refactor: 컨트롤러 DTO 네이밍 변경 및 반영
